requirements.txt
# ---- PyTorch (stable & compatible with YOLOv8) ----
torch==2.3.1
torchvision==0.18.1
torchaudio==2.3.1

# ---- YOLO ----
ultralytics==8.2.48

# ---- API ----
fastapi
uvicorn
python-multipart

# ---- Image & Utils ----
opencv-python-headless
numpy
pillow
tqdm
pyyaml
